scraping/requirements.txt
pandas
numpy
chess
pymysql
python-dotenv
black
requests
pyppeteer
country_converter
